Rename consensus_params.rs to chain_params.rs
We decided that ChainParams + ValidatorSet = ConsensusParams.
